+++ compiled/src/event-listener.decorator.js
@@ -1,5 +1,6 @@
 "use strict";
 Object.defineProperty(exports, "__esModule", { value: true });
+var event_manager_1 = require("./event-manager");
 function getClassPropertyName(args, instance) {
     if (!args.eventClass)
         return null;
@@ -16,13 +17,37 @@ function getEventClass(args, classInstance) {
     var propertyName = getClassPropertyName(args, classInstance);
     var instance = (propertyName) ? classInstance[propertyName] : classInstance;
     if (!instance.on) {
-        if (instance.events && instance.events.on) {
-            return instance.events;
+        var propNames = Object.getOwnPropertyNames(instance);
+        for (var _i = 0, propNames_1 = propNames; _i < propNames_1.length; _i++) {
+            var propName = propNames_1[_i];
+            if (instance[propName] && instance[propName].constructor
+                && instance[propName].constructor.name === event_manager_1.EventManager.name) {
+                return instance[propName];
+            }
         }
         throw new Error('@EventListener: Class ' + instance.constructor.name
             + ' must extend EventManager');
     }
     return instance;
+}
+function validateFunctionsExist(args, target) {
+    if (args.initFn || args.destroyFn) {
+        if (!args.initFn || !args.destroyFn) {
+            throw new Error('@EventListener: both initFn and destroyFn have to be '
+                + 'set together');
+        }
+    }
+    else {
+        if (!target.constructor.prototype.ngOnInit) {
+            throw new Error('@EventListener: Class ' + target.constructor.name +
+                ' must implement ngOnInit method');
+        }
+        if (!target.constructor.prototype.ngOnDestroy) {
+            throw new Error('@EventListener: Class ' + target.constructor.name +
+                ' must implement ngOnDestroy method');
+        }
+    }
+    return true;
 }
 function EventListener(arg1, arg2) {
     return function (target, propertyKey, descriptor) {
@@ -36,6 +61,7 @@ function EventListener(arg1, arg2) {
         }
         var params = (typeof arg1 === 'string')
             ? { eventName: arg1, eventClass: arg2 } : arg1;
+        validateFunctionsExist(params, target);
         var initFnName = (params.initFn || 'ngOnInit');
         var destroyFnName = (params.destroyFn || 'ngOnDestroy');
         var eventClassInstance = null;