+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# 1.3.1
+
+__Important:__ When using with Angular, your components must implement ngOnInit and ngOnDestroy functions. See fixed issue below.
+
+### Fixed:
+
+* Fixed issue with Angular life cycle hooks not working with Decorators [16023](https://github.com/angular/angular/issues/16023), [31495](https://github.com/angular/angular/issues/31495).
+
 # 1.3.0
 
 ### Added: